tests: Move util.py -> utils.py
